requirements.txt
mysql-connector-python==8.0.28
virtualenv==20.13.0